pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "datetime>=5.5",
    "mediapipe>=0.10.21",
    "mss>=10.0.0",
    "numpy>=1.26.4",
    "openai>=1.95.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
]
